[Graphics] - Reworking methods to use spans instead of arrays. [Imaging] - Updating image functionality to use new pointer functionality in the core API. [Imaging] - Added block compression functionality for compressing/decompressing image data via the BCn formats. [Imaging] - Moved image editing functionality into standard fluent interfaces. Changed from using an extension class to using the image concrete class in order to simplify the functionality. [Imaging] - Changed image editing to work as a command list which is executed at the end when EndUpdate is executed. [Imaging] - Removed CopyFrom/CopyTo methods and replaced them with a single Copy method which copies the image passed in via the parameter. [FileSystem] - Created general base abstract class for the file system provider and moved the folder file system provider into an internal class. [FileSystem] - Added read only flag to provider base class. [FileSystem] - Updated writer code to disallow updating files/directories that are mounted on a real only provider. [FileSystem] - Updated provider factory to load assemblies with provider plug ins instead of just passing a plug in name. [FileSystem] - General clean up. [Input] - Reworked code to be more SOLID compliant. [Input] - Removed dependency on System.Drawing for rectangles/points, uses SharpDX rectangles and points. [Input] - Fixed up internal raw input dispatching to be more efficient when passing data and to use the new pointer wrapper. [Input] - Update factory methods to simplify loading joystick driver plug ins. [Windows] - Updated console log to use updated logging base class. [Windows] - Fixed bug in code that retrieves shell icons where COM interface was not explicitly released when not needed any more. [Windows] - Updated dialogs class to use IWin32Window as parent instead of Form. [Windows] - Deleted old GorgonOverlayPanel, GorgonProgressScreenPanel, and GorgonWaitScreenPanel as they are outdated, slow, and not very good. [Windows] - Made FormWait, and FormProgress internal - use overlay objects instead. [Windows] - Added new overlay objects to display overlays on forms/controls. Currently has 3 types of overlay: Overlay - general translucent overlay, Wait Overlay - translucent overlay with wait message, and Progress Overlay - translucent overlay with progress meter. [Core] - Added subtraction operator to pointer types to return the number of bytes between two pointers. [Core] - Updated internal plug in registration code to only work with netstandard2.0. This is to handle future update to .NET 5.0 which uses a nuget package for this. [Core] - Reworked how logging objects are created by using a base class for logging, and concrete types for logging via console or text file. [Core] - Added depth verse tree traversal extension method. [Core] - Added methods to GorgonArray that will return read only Span<T> and read only Memory<T> types. Useful for slicing the array. [Core] - Changed extension methods for Vectors and Matrices to use System.Numerics instead of SharpDX. The system numerics, despite not having ref/out parameters still performed better than the SharpDX counterparts (due to SIMD no doubt).  Extension methods to pass information by in/ref/out have been created to allow taking advantage of that optimization. [Core] - Added a MatrixFactory static class to create various specific Matrix types using in/ref/out. [Core] - Extensive rework of native buffer class, and added a new struct type for accessing native memory as a pointer. [Core] - Updated GorgonBinaryReader/Writer to use spans and use new native pointer struct. [Core] - Removed unnecessary code. [Examples] - Updated file system examples to use updated factory code. [Examples] - Updated all examples to use updated input API. [Examples] - Fixed up memory leak in input device example. [Examples] - Added example for overlay functionality, and folder browser functionality from Gorgon.Windows. [Resources] - Added resource directory hierarchy for folder browser example so users can play with the control without wrecking their own file system.
Added sub module for the Gorgon specific fork of BCnEncoder.NET. This is now referenced by the Gorgon.Imaging project. This means that those who grab the code from Github will need to pull the submodules as well.

Configuring repo to ignore dirty changes in submodule.

General code clean up.

+++ Gorgon/Gorgon.FileSystem/GorgonFileSystemWriter.cs
@@ -499,15 +499,18 @@ namespace Gorgon.IO
             // If the file doesn't exist, then we don't need to actually delete it.
             if (File.Exists(physicalPath))
             {
-                if (_deleteAction == null)
+                if (!file.MountPoint.Provider.IsReadOnly)
                 {
-                    File.Delete(physicalPath);
-                }
-                else
-                {
-                    if (!_deleteAction(physicalPath))
+                    if (_deleteAction == null)
                     {
-                        return false;
+                        File.Delete(physicalPath);
+                    }
+                    else
+                    {
+                        if (!_deleteAction(physicalPath))
+                        {
+                            return false;
+                        }
                     }
                 }
 
@@ -773,15 +776,18 @@ namespace Gorgon.IO
 
                 if (Directory.Exists(physicalPath))
                 {
-                    if (_deleteAction == null)
+                    if (!dir.MountPoint.Provider.IsReadOnly)
                     {
-                        Directory.Delete(physicalPath);
-                    }
-                    else
-                    {
-                        if (!_deleteAction(physicalPath))
+                        if (_deleteAction == null)
                         {
-                            continue;
+                            Directory.Delete(physicalPath);
+                        }
+                        else
+                        {
+                            if (!_deleteAction(physicalPath))
+                            {
+                                continue;
+                            }
                         }
                     }
 
@@ -998,12 +1004,15 @@ namespace Gorgon.IO
             string physicalPath = GetWriteFilePath(file.Directory.FullPath, file.Name);
             string newPhysicalPath = GetWriteFilePath(file.Directory.FullPath, newName);
 
-            if (File.Exists(physicalPath))
-            {
-                File.Move(physicalPath, newPhysicalPath);
-            }
+            if (!file.MountPoint.Provider.IsReadOnly)
+            {            
+                if (File.Exists(physicalPath))
+                {
+                    File.Move(physicalPath, newPhysicalPath);
+                }
 
-            _notifier.NotifyFileRenamed(_mountPoint, oldPath, new PhysicalFileInfo(_mountPoint, newPhysicalPath));
+                _notifier.NotifyFileRenamed(_mountPoint, oldPath, new PhysicalFileInfo(_mountPoint, newPhysicalPath));
+            }
 
             EventHandler<VirtualFileRenamedEventArgs> handler = VirtualFileRenamed;
             handler?.Invoke(this, new VirtualFileRenamedEventArgs(file, oldName));
@@ -1080,13 +1089,16 @@ namespace Gorgon.IO
             string physicalPath = GetWriteDirectoryPath(directory.FullPath);
             string newPhysicalPath = GetWriteDirectoryPath(directory.Parent.FullPath + newName);
 
-            if (Directory.Exists(physicalPath))
+            if (!directory.MountPoint.Provider.IsReadOnly)
             {
-                // .NET has no facility (at least none that I'm aware of) to rename a directory, so we move it instead, hence why we need the full path for the new name.
-                Directory.Move(physicalPath, newPhysicalPath);
-            }
+                if (Directory.Exists(physicalPath))
+                {
+                    // .NET has no facility (at least none that I'm aware of) to rename a directory, so we move it instead, hence why we need the full path for the new name.
+                    Directory.Move(physicalPath, newPhysicalPath);
+                }
 
-            _notifier.NotifyDirectoryRenamed(_mountPoint, oldPath, newPhysicalPath, newName);
+                _notifier.NotifyDirectoryRenamed(_mountPoint, oldPath, newPhysicalPath, newName);
+            }
 
             EventHandler<VirtualDirectoryRenamedEventArgs> handler = VirtualDirectoryRenamed;
             handler?.Invoke(this, new VirtualDirectoryRenamedEventArgs(directory, oldName));
@@ -1285,7 +1297,7 @@ namespace Gorgon.IO
                         BlockCopyFile(srcFile, destFilePath, progressCallback, cancelToken);                        
 
                         // Delete the source file, we won't need it anymore.                        
-                        if (File.Exists(srcFile.PhysicalFile.FullPath))
+                        if ((File.Exists(srcFile.PhysicalFile.FullPath)) && (!_mountPoint.Provider.IsReadOnly))
                         {
                             File.Delete(srcFile.PhysicalFile.FullPath);
                         }
@@ -1309,7 +1321,7 @@ namespace Gorgon.IO
                     }
 
                     string physicalPath = _mountPoint.Provider.MapToPhysicalPath(srcDir.FullPath, _mountPoint);
-                    if (Directory.Exists(physicalPath))
+                    if ((Directory.Exists(physicalPath)) && (!_mountPoint.Provider.IsReadOnly))
                     {
                         Directory.Delete(physicalPath);                        
                     }
@@ -1660,7 +1672,7 @@ namespace Gorgon.IO
                     BlockCopyFile(srcFile, destFilePath, progressCallback, cancelToken);
 
                     // Delete the source file, we won't need it anymore.                        
-                    if (File.Exists(srcFile.PhysicalFile.FullPath))
+                    if ((File.Exists(srcFile.PhysicalFile.FullPath)) && (!_mountPoint.Provider.IsReadOnly))
                     {
                         File.Delete(srcFile.PhysicalFile.FullPath);
                     }
@@ -2502,8 +2514,8 @@ namespace Gorgon.IO
                 throw new ArgumentEmptyException(nameof(writeLocation));
             }
 
-            _deleteAction = deleteAction;
             FileSystem = fileSystem ?? throw new ArgumentNullException(nameof(fileSystem));
+            _deleteAction = deleteAction;
             _notifier = notifier ?? throw new ArgumentNullException(nameof(notifier));
             WriteLocation = writeLocation.FormatDirectory(Path.DirectorySeparatorChar);
             _mountPoint = new GorgonFileSystemMountPoint(fileSystem.DefaultProvider, WriteLocation, "/");
